+++ src/language.ts
@@ -115,7 +115,7 @@ export const ISOLanguage = {
     name: "Amharic",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -163,7 +163,7 @@ export const ISOLanguage = {
     name: "Basque",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -187,7 +187,7 @@ export const ISOLanguage = {
     name: "Bengali",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -283,7 +283,7 @@ export const ISOLanguage = {
     name: "Corsican",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -463,7 +463,7 @@ export const ISOLanguage = {
     name: "Hindi",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -499,7 +499,7 @@ export const ISOLanguage = {
     name: "Indonesian",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -643,7 +643,7 @@ export const ISOLanguage = {
     name: "Kurdish",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -715,7 +715,7 @@ export const ISOLanguage = {
     name: "Malayalam",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -751,7 +751,7 @@ export const ISOLanguage = {
     name: "Occitan",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -763,7 +763,7 @@ export const ISOLanguage = {
     name: "Persian",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -799,7 +799,7 @@ export const ISOLanguage = {
     name: "Punjabi",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -943,7 +943,7 @@ export const ISOLanguage = {
     name: "Tamil",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -955,7 +955,7 @@ export const ISOLanguage = {
     name: "Telugu",
     latin: false,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -1003,7 +1003,7 @@ export const ISOLanguage = {
     name: "Vietnamese (Latin script)",
     latin: true,
     isMode: false,
-    geocoding: false,
+    geocoding: true,
   } as LanguageInfo,
 
   /**
@@ -1034,7 +1034,7 @@ export const Language = {
  */
 export function getLanguageInfoFromKey(
   languageKey: string,
-  languageDictionary: { [k: string]: LanguageInfo } = Language,
+  languageDictionary: { [k: string]: LanguageInfo } = Language
 ): LanguageInfo | null {
   if (languageKey in languageDictionary) {
     return languageKey[languageKey];
@@ -1049,7 +1049,7 @@ export function getLanguageInfoFromKey(
  */
 export function getLanguageInfoFromCode(
   languageCode: string,
-  languageDictionary: { [k: string]: LanguageInfo } = Language,
+  languageDictionary: { [k: string]: LanguageInfo } = Language
 ): LanguageInfo | null {
   for (const lang of Object.values(languageDictionary)) {
     if (lang.code === languageCode) {
@@ -1067,7 +1067,7 @@ export function getLanguageInfoFromCode(
  */
 export function getLanguageInfoFromFlag(
   languageFlag: string,
-  languageDictionary: { [k: string]: LanguageInfo } = Language,
+  languageDictionary: { [k: string]: LanguageInfo } = Language
 ): LanguageInfo | null {
   for (const lang of Object.values(languageDictionary)) {
     if (lang.flag === languageFlag) {
@@ -1088,7 +1088,7 @@ export function getAutoLanguage(): LanguageInfo {
   }
 
   const canditatelangs = Array.from(
-    new Set(navigator.languages.map((l) => l.split("-")[0])),
+    new Set(navigator.languages.map((l) => l.split("-")[0]))
   )
     .map((code) => getLanguageInfoFromCode(code))
     .filter((li) => li);
@@ -1120,7 +1120,7 @@ export function isLanguageInfo(obj: unknown): obj is LanguageInfo {
  */
 export function toLanguageInfo(
   lang: LanguageInfo | string,
-  languageDictionary: { [k: string]: LanguageInfo } = Language,
+  languageDictionary: { [k: string]: LanguageInfo } = Language
 ): LanguageInfo | null {
   // Could be directly an object of type LanguageInfo
   if (isLanguageInfo(lang)) {
@@ -1149,7 +1149,7 @@ export function toLanguageInfo(
 export function areSameLanguages(
   langA: string | LanguageInfo,
   langB: string | LanguageInfo,
-  languageDictionary: { [k: string]: LanguageInfo } = Language,
+  languageDictionary: { [k: string]: LanguageInfo } = Language
 ): boolean {
   const langAObj = toLanguageInfo(langA, languageDictionary);
   const langBObj = toLanguageInfo(langB, languageDictionary);